streamline refactor (2)
