--- examples/data/product_specs.json
@@ -1,261 +1,0 @@
-[
-  {
-    "id": "head1",
-    "type": "heading_1",
-    "content": "Mobile App Product Specification",
-    "parent_id": null,
-    "has_children": true
-  },
-  {
-    "id": "para1",
-    "type": "paragraph",
-    "content": "Version 1.0 | Last Updated: March 25, 2023",
-    "parent_id": "head1",
-    "has_children": false
-  },
-  {
-    "id": "head2_1",
-    "type": "heading_2",
-    "content": "Product Overview",
-    "parent_id": "head1",
-    "has_children": true
-  },
-  {
-    "id": "para2",
-    "type": "paragraph",
-    "content": "Our mobile application aims to simplify task management for busy professionals with a focus on simplicity, speed, and smart organization features.",
-    "parent_id": "head2_1",
-    "has_children": false
-  },
-  {
-    "id": "head2_2",
-    "type": "heading_2",
-    "content": "Target Audience",
-    "parent_id": "head1",
-    "has_children": true
-  },
-  {
-    "id": "bullet1",
-    "type": "bulleted_list_item",
-    "content": "Professionals aged 25-45",
-    "parent_id": "head2_2",
-    "has_children": false
-  },
-  {
-    "id": "bullet2",
-    "type": "bulleted_list_item",
-    "content": "Small business owners",
-    "parent_id": "head2_2",
-    "has_children": false
-  },
-  {
-    "id": "bullet3",
-    "type": "bulleted_list_item",
-    "content": "Freelancers and consultants",
-    "parent_id": "head2_2",
-    "has_children": false
-  },
-  {
-    "id": "bullet4",
-    "type": "bulleted_list_item",
-    "content": "Team managers",
-    "parent_id": "head2_2",
-    "has_children": false
-  },
-  {
-    "id": "head2_3",
-    "type": "heading_2",
-    "content": "Core Features",
-    "parent_id": "head1",
-    "has_children": true
-  },
-  {
-    "id": "head3_1",
-    "type": "heading_3",
-    "content": "Task Management",
-    "parent_id": "head2_3",
-    "has_children": true
-  },
-  {
-    "id": "bullet5",
-    "type": "bulleted_list_item",
-    "content": "Create, edit, and delete tasks",
-    "parent_id": "head3_1",
-    "has_children": false
-  },
-  {
-    "id": "bullet6",
-    "type": "bulleted_list_item",
-    "content": "Set due dates and reminders",
-    "parent_id": "head3_1",
-    "has_children": false
-  },
-  {
-    "id": "bullet7",
-    "type": "bulleted_list_item",
-    "content": "Priority levels (High, Medium, Low)",
-    "parent_id": "head3_1",
-    "has_children": false
-  },
-  {
-    "id": "bullet8",
-    "type": "bulleted_list_item",
-    "content": "Task categories and labels",
-    "parent_id": "head3_1",
-    "has_children": false
-  },
-  {
-    "id": "head3_2",
-    "type": "heading_3",
-    "content": "Smart Organization",
-    "parent_id": "head2_3",
-    "has_children": true
-  },
-  {
-    "id": "bullet9",
-    "type": "bulleted_list_item",
-    "content": "AI-powered task categorization",
-    "parent_id": "head3_2",
-    "has_children": false
-  },
-  {
-    "id": "bullet10",
-    "type": "bulleted_list_item",
-    "content": "Natural language processing for task creation",
-    "parent_id": "head3_2",
-    "has_children": true
-  },
-  {
-    "id": "bullet10_1",
-    "type": "bulleted_list_item",
-    "content": "\"Meeting with John tomorrow at 3pm\" → Creates task with date and time",
-    "parent_id": "bullet10",
-    "has_children": false
-  },
-  {
-    "id": "bullet10_2",
-    "type": "bulleted_list_item",
-    "content": "\"High priority: Finish report by Friday\" → Sets priority and deadline",
-    "parent_id": "bullet10",
-    "has_children": false
-  },
-  {
-    "id": "bullet11",
-    "type": "bulleted_list_item",
-    "content": "Automated daily agenda",
-    "parent_id": "head3_2",
-    "has_children": false
-  },
-  {
-    "id": "head3_3",
-    "type": "heading_3",
-    "content": "Collaboration",
-    "parent_id": "head2_3",
-    "has_children": true
-  },
-  {
-    "id": "bullet12",
-    "type": "bulleted_list_item",
-    "content": "Share tasks and lists with team members",
-    "parent_id": "head3_3",
-    "has_children": false
-  },
-  {
-    "id": "bullet13",
-    "type": "bulleted_list_item",
-    "content": "Assign tasks to others",
-    "parent_id": "head3_3",
-    "has_children": false
-  },
-  {
-    "id": "bullet14",
-    "type": "bulleted_list_item",
-    "content": "Real-time updates and notifications",
-    "parent_id": "head3_3",
-    "has_children": false
-  },
-  {
-    "id": "bullet15",
-    "type": "bulleted_list_item",
-    "content": "Comment on tasks",
-    "parent_id": "head3_3",
-    "has_children": false
-  },
-  {
-    "id": "head2_4",
-    "type": "heading_2",
-    "content": "Technical Requirements",
-    "parent_id": "head1",
-    "has_children": true
-  },
-  {
-    "id": "code1",
-    "type": "code",
-    "content": "// Minimum OS versions\niOS: 14.0+\nAndroid: 8.0+\n\n// Backend Requirements\n- Node.js API server\n- MongoDB database\n- Redis for caching\n- AWS S3 for file storage",
-    "parent_id": "head2_4",
-    "has_children": false
-  },
-  {
-    "id": "head2_5",
-    "type": "heading_2",
-    "content": "Future Enhancements (v2.0)",
-    "parent_id": "head1",
-    "has_children": true
-  },
-  {
-    "id": "para3",
-    "type": "paragraph",
-    "content": "The following features are planned for the second major release:",
-    "parent_id": "head2_5",
-    "has_children": false
-  },
-  {
-    "id": "num1",
-    "type": "numbered_list_item",
-    "content": "Calendar integration with Google and Outlook",
-    "parent_id": "head2_5",
-    "has_children": false
-  },
-  {
-    "id": "num2",
-    "type": "numbered_list_item",
-    "content": "Advanced analytics dashboard",
-    "parent_id": "head2_5",
-    "has_children": false
-  },
-  {
-    "id": "num3",
-    "type": "numbered_list_item",
-    "content": "Offline mode with full functionality",
-    "parent_id": "head2_5",
-    "has_children": false
-  },
-  {
-    "id": "num4",
-    "type": "numbered_list_item",
-    "content": "Voice commands for hands-free operation",
-    "parent_id": "head2_5",
-    "has_children": false
-  },
-  {
-    "id": "head2_6",
-    "type": "heading_2",
-    "content": "Timeline",
-    "parent_id": "head1",
-    "has_children": true
-  },
-  {
-    "id": "quote1",
-    "type": "quote",
-    "content": "All dates are subject to change based on development progress and stakeholder feedback.",
-    "parent_id": "head2_6",
-    "has_children": false
-  },
-  {
-    "id": "para4",
-    "type": "paragraph",
-    "content": "Alpha Release: June 2023\nBeta Testing: August 2023\nPublic Launch: October 2023",
-    "parent_id": "head2_6",
-    "has_children": false
-  }
-]
